[ExecutiveButtons, OriginalData] Fix update and delete buttons, with intermediary buttons, warnings, and resetting state

+++ src/components/elements/predictions/OriginalData.js
@@ -19,7 +19,7 @@ function OriginalData(props) {
         if (submitted) {
             setSubmitted(false)
         }
-    })
+    }, [submitted])
 
     let source = props.source
     const user = props.user
@@ -61,53 +61,104 @@ function OriginalData(props) {
         setDataSet(e.target.value)
     }
 
+    const handleUndoSubmit = (e) => {
+        e.preventDefault()
+        const submitButton = document.getElementById('submit-button')
+        const undoButton = document.getElementById('undo-submit-button')
+        const warning = document.getElementById('submit-warning')
+        const deleteButton = document.getElementById('delete-button')
+        submitButton.innerText = 'Update Set'
+        undoButton.style.display = 'none'
+        warning.style.display = 'none'
+        deleteButton.style.display = 'block'
+        setTitle(props.title)
+        setIndependent(props.independent)
+        setDependent(props.dependent)
+        setPrecision(props.precision)
+        setDataSet(props.dataSet)
+        window.scrollTo(0, 0)
+    }
+    
+    const handleUndoDelete = (e) => {
+        e.preventDefault()
+        const deleteButton = document.getElementById('delete-button')
+        const undoButton = document.getElementById('undo-delete-button')
+        const warning = document.getElementById('delete-warning')
+        const submitButton = document.getElementById('submit-button')
+        deleteButton.innerText = 'Delete Set'
+        undoButton.style.display = 'none'
+        warning.style.display = 'none'
+        submitButton.style.display = 'block'
+        window.scrollTo(0, 0)
+    }
+
     const handleSubmit = async (e) => {
         e.preventDefault()
-        if (title === '') {
-            alert('You must give this data set a title')
-        } else if (independent === '') {
-            alert('You must identify the independent variable of this data set')
-        } else if (dependent === '') {
-            alert('You must identify the dependent variable of this data set')
-        } else if (precision < 1) {
-            alert('Precision must be a positive integer')
-        } else if (dataSet === '') {
-            alert('You must provide a data set')
+        const submitButton = document.getElementById('submit-button')
+        const undoButton = document.getElementById('undo-submit-button')
+        const warning = document.getElementById('submit-warning')
+        const deleteButton = document.getElementById('delete-button')
+
+        if (submitButton.innerText === 'Update Set') {
+            submitButton.innerText = 'Update Set with New Data'
+            undoButton.style.display = 'block'
+            warning.style.display = 'block'
+            deleteButton.style.display = 'none'
         } else {
-            const cleanedDataSet = CleanCollection(dataSet)
-            const parsedDataSet = JSON.parse(cleanedDataSet)
-            if (parsedDataSet.length < 10) {
-                alert('You must include at least 10 points in your data set')
+            if (title === '') {
+                alert('You must give this data set a title')
+            } else if (independent === '') {
+                alert('You must identify the independent variable of this data set')
+            } else if (dependent === '') {
+                alert('You must identify the dependent variable of this data set')
+            } else if (precision < 1) {
+                alert('Precision must be a positive integer')
+            } else if (dataSet === '') {
+                alert('You must provide a data set')
             } else {
-                const submission = {
-                    title: title,
-                    independent: independent,
-                    dependent: dependent,
-                    precision: parseInt(precision),
-                    dataSet: parsedDataSet
-                }
-                if (!stored) {
-                    try {
-                        const predictions = await axios.post(REACT_APP_SERVER_URL + 'api',
-                            submission
-                        )
-                        setModels(predictions.data.regressions)
-                        setSubmitted(true)
-                    } catch (error) {
-                        alert(error)
-                    }
+                const cleanedDataSet = CleanCollection(dataSet)
+                const parsedDataSet = JSON.parse(cleanedDataSet)
+                if (parsedDataSet.length < 10) {
+                    alert('You must include at least 10 points in your data set')
                 } else {
-                    try {
-                        await axios.delete(REACT_APP_SERVER_URL + 'predictions/' + source)
-                        const predictions = await axios.post(REACT_APP_SERVER_URL + 'api', submission)
-                        source = predictions.data.regressions.source
-                        await axios.post(REACT_APP_SERVER_URL + 'predictions/' + user.id, { source })
-                        await axios.put(REACT_APP_SERVER_URL + 'predictions/' + source + '/favorite', {favorite: opinions.favorite})
-                        await axios.put(REACT_APP_SERVER_URL + 'predictions/' + source + '/note', {note: opinions.note})
-                        setModels(predictions.data.regressions)
-                        setSubmitted(true)
-                    } catch(error) {
-                        alert(error)
+                    const submission = {
+                        title: title,
+                        independent: independent,
+                        dependent: dependent,
+                        precision: parseInt(precision),
+                        dataSet: parsedDataSet
+                    }
+                    if (!stored) {
+                        try {
+                            const predictions = await axios.post(REACT_APP_SERVER_URL + 'api',
+                                submission
+                            )
+                            submitButton.innerText = 'Update Set'
+                            deleteButton.style.display = 'block'
+                            undoButton.style.display = 'none'
+                            warning.style.display = 'none'
+                            setModels(predictions.data.regressions)
+                            setSubmitted(true)
+                        } catch (error) {
+                            alert(error)
+                        }
+                    } else {
+                        try {
+                            await axios.delete(REACT_APP_SERVER_URL + 'predictions/' + source)
+                            const predictions = await axios.post(REACT_APP_SERVER_URL + 'api', submission)
+                            source = predictions.data.regressions.source
+                            await axios.post(REACT_APP_SERVER_URL + 'predictions/' + user.id, { source })
+                            await axios.put(REACT_APP_SERVER_URL + 'predictions/' + source + '/favorite', {favorite: opinions.favorite})
+                            await axios.put(REACT_APP_SERVER_URL + 'predictions/' + source + '/note', {note: opinions.note})
+                            submitButton.innerText = 'Update Set'
+                            deleteButton.style.display = 'block'
+                            undoButton.style.display = 'none'
+                            warning.style.display = 'none'
+                            setModels(predictions.data.regressions)
+                            setSubmitted(true)
+                        } catch(error) {
+                            alert(error)
+                        }
                     }
                 }
             }
@@ -115,12 +166,28 @@ function OriginalData(props) {
     }
 
     const handleDelete = async (e) => {
-        try {
-            e.preventDefault()
-            await axios.delete(REACT_APP_SERVER_URL + 'predictions/' + source)
-            setDeleted(true)
-        } catch(error) {
-            alert(error)
+        e.preventDefault()
+        const deleteButton = document.getElementById('delete-button')
+        const undoButton = document.getElementById('undo-delete-button')
+        const warning = document.getElementById('delete-warning')
+        const submitButton = document.getElementById('submit-button')
+
+        if (deleteButton.innerText === 'Delete Set') {
+            deleteButton.innerText = 'Yes, Delete Set'
+            undoButton.style.display = 'block'
+            warning.style.display = 'block'
+            submitButton.style.display = 'none'
+        } else {
+            try {
+                await axios.delete(REACT_APP_SERVER_URL + 'predictions/' + source)
+                deleteButton.innerText = 'Delete Set'
+                undoButton.style.display = 'none'
+                warning.style.display = 'none'
+                submitButton.style.display = 'block'
+                setDeleted(true)
+            } catch(error) {
+                alert(error)
+            }
         }
     }
 
@@ -141,6 +208,8 @@ function OriginalData(props) {
                 handleDataSet={handleDataSet}
                 stored={props.stored}
                 button={buttonText}
+                handleUndoSubmit={handleUndoSubmit}
+                handleUndoDelete={handleUndoDelete}
             />
         )
     } else if (submitted) {
